Refactor the Credential Re-Up Code.
Rather than waiting an explicit number of seconds, check every 15
seconds if the creds have expired or not (without an API call). When the
creds are 10m away from expiring, go and actively get new ones.

This makes it easier for us to handle network connection blips (closed
laptop, etc) and reduces the complexity for the user.

Also in here is a slight change to the way the Session object works --
treating it more like a real object (with properties, etc).