Support --account and --role to choose the account or role when multiple are presented.
Supplying the --account and/or --role parameters filters the list when
multiple accounts are returned. If the list filters to a single item,
that one used without prompt. If multiple matches occur, a reduced list
is presented.

The primary difference is a change from triggering the call to
handle_multiple_roles on error to checking for multiple accounts up front.

The list of accounts and roles is fetched and filtered. If the filtering
results in multiple accounts/roles, the filtered list is presented for
selection. If only one item is returned, it proceeds with that selection.

Testing Done:
* 100% test coverage
* attempted with variatons against both a single- and multi-
  account/role okta config