Fix issue where sorting misaligns the list indexes.
fixes #71

Sorting is applied after the list index positions
are assigned to the role dicts. The roldIdx value
is then used to choose the role arn by position, but
the position has changed.

This change moves the index assignment to after list
sort. A test was updated to verify indexes after sorting.